doctors/requirements.txt
Django==3.1.7
dj-database-url==0.5.0
psycopg2==2.8.6
djangorestframework==3.12.2
requests==2.25.1
PyJWT==2.0.1